- Removed a small memory leak dealing with loaded <garbage> objects.     [FB6 fix] - Changed the exit cost deduction later in the function to prevent     a player from getting unfairly charged. [FB6 fix] Does anyone     even USE in-server pennies for anything anymore? :P - Added a .debug/lastcrashtime prop to get written to programs when they     crash. It's in a human readable time format rathern than a systime.     [FB6 idea] - Fixed memory leak in ARRAY_FMTSTRINGS. [FB6 bug] - Fixed FMTSTRING to not accept negative dynamic precision strings.     [FB6 bug] - Fixed an error with FMTSTRING accepting a dynamic field width of -1.     [FB6 bug] - Added Standard ANSI awareness to the FMTSTRING and ARRAY_FMTSTRINGS     prims. [FB6 improvement] - Made it so that scoped variables will have their name shown in debug     mode. [FB6 idea] - Added ABS prim. [FB6 idea] - Added SIGN prim. [FB6 idea. Kind of pointless. :P] - Fixed a crasher when having nameless variables in function headers.     [FB6 fix] - Added DESCRBUFSIZE prim. [FB6 prim] - Added }TELL in-server $define. [FB6 idea] - Fixed potential crasher in ARRAY_FILTER_PROP. [FB6 fix] - Fixed double-freeing in ARRAY_FMTSTRINGS. [FB6 fix]
Whee! FB6 catchup work!
- Akari

+++ protomuck/src/array.c
@@ -527,7 +527,7 @@ new_array_packed(int size)
 	}
 
 	new2 = new_array();
-	new2->items = size;
+	new2->items = (short int)size;
 	new2->type = ARRAY_PACKED;
 	if (size < 1)
 		size = 1;
